# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.TripEmailItem:
    description: An email associated with a trip
    properties:
      date:
        type: string
      filename:
        type: string
      from:
        type: string
      message_id:
        type: string
      subject:
        type: string
      tags:
        items:
          type: string
        type: array
      thread_id:
        type: string
    type: object
  handlers.associateTripEmailResponse:
    properties:
      message_id:
        example: <abc123@gmail.com>
        type: string
      tags:
        example:
        - inbox
        - trip:2026-cruise
        - travel
        items:
          type: string
        type: array
      trip_id:
        example: 2026-cruise
        type: string
    type: object
  handlers.createTripRequest:
    properties:
      date_range:
        type: string
      name:
        type: string
    type: object
  handlers.tripEmailsResponse:
    properties:
      emails:
        items:
          $ref: '#/definitions/handlers.TripEmailItem'
        type: array
      trip_id:
        example: 2026-cruise
        type: string
    type: object
  handlers.tripResponse:
    properties:
      confirmed_count:
        type: integer
      created_at:
        type: string
      date_range:
        type: string
      email_count:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  notmuch.EmailResult:
    description: Email search result
    properties:
      body_preview:
        type: string
      category:
        type: string
      date:
        example: "2023-01-01T12:00:00Z"
        type: string
      filename:
        example: /path/to/email.eml
        type: string
      from:
        example: sender@example.com
        type: string
      message_id:
        example: <12345@example.com>
        type: string
      subject:
        example: Flight Confirmation
        type: string
      tags:
        example:
        - travel
        - flight
        items:
          type: string
        type: array
      thread_id:
        example: thread123
        type: string
      trip_id:
        type: string
    type: object
  notmuch.SearchResults:
    description: Search results containing matching emails
    properties:
      count:
        example: 42
        type: integer
      query:
        example: subject:flight
        type: string
      results:
        items:
          $ref: '#/definitions/notmuch.EmailResult'
        type: array
    type: object
host: localhost:8181
info:
  contact:
    email: support@example.com
    name: API Support
  description: A self-hosted travel plan aggregator that searches through emails
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Voyage API
  version: "1.0"
paths:
  /email/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single email by its message ID
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notmuch.EmailResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get email by ID
      tags:
      - email
  /email/{id}/tags/{tag}:
    post:
      consumes:
      - application/json
      description: Add a tag to an email by its message ID
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag to add
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notmuch.EmailResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Tag an email
      tags:
      - email
  /email/{id}/trip/{tripId}:
    post:
      consumes:
      - application/json
      description: Remove any existing trip:* tags and apply trip:{tripId} to the
        email
      parameters:
      - description: Email message ID
        in: path
        name: id
        required: true
        type: string
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.associateTripEmailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Associate an email with a trip
      tags:
      - email
      - trips
  /health:
    get:
      consumes:
      - application/json
      description: Get the health status of the API and database connection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health check endpoint
      tags:
      - health
  /search:
    get:
      consumes:
      - application/json
      description: Search for emails using notmuch query
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: "50"
        description: Result limit
        in: query
        name: limit
        type: string
      - default: newest_first
        description: Sort order (oldest_first, newest_first)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notmuch.SearchResults'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search emails
      tags:
      - search
  /trips:
    get:
      consumes:
      - application/json
      description: Get a list of all trips
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.tripResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all trips
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Create a trip with a name and date range
      parameters:
      - description: Trip to create
        in: body
        name: trip
        required: true
        schema:
          $ref: '#/definitions/handlers.createTripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.tripResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new trip
      tags:
      - trips
  /trips/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a trip by id
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a trip
      tags:
      - trips
  /trips/{id}/emails:
    get:
      consumes:
      - application/json
      description: Get all emails tagged with a specific trip, sorted chronologically
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.tripEmailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List emails for a trip
      tags:
      - email
      - trips
swagger: "2.0"